backend/requirements.txt
flask
flask-cors
psycopg2-binary
pdfplumber
python-docx
spacy
python-magic
Werkzeug
